remove vscode
